[[ app ]] Update the app jni and java source files to support latest platforms
This patch updates both the jni and java source files so that they will work on all
four standard architectures, and with android platform 33.

The source has been updated to use a long to hold the native version of the
synthesis class, to enable functioning on 64-bit platforms.

The source has been updated so that the data storage folder is computed on startup
and stored in a static variable as in the newer platform api context is required
to compute it.